requirements.txt
fastapi
uvicorn[standard]
gitpython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
tqdm>=4.64.0
onnx==1.12.0
protobuf<=3.20.0
paddlepaddle==2.4.1
paddleocr
opencv-python